+++ src/components/projects/ProjectTable.tsx
@@ -12,6 +12,7 @@ import deleteData from "@/hooks/deleteData";
 import { useEffect, useState } from "react";
 import getData from "@/hooks/getData";
 import Image from "next/image";
+import Swal from "sweetalert2";
 type ProjectItem = {
   _id: string;
   title: string;
@@ -48,20 +49,37 @@ const ProjectTable = () => {
 
     fetchData();
   }, []);
+
   const handleDelete = async (id: string) => {
-    const res = await deleteData({
-      url: `/dashboard/projects/api?id=${id}`,
-      setLoader,
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to undo this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "Cancel",
     });
 
-    if (!res.error) {
-      setProjects((prev) => prev.filter((blog) => blog._id !== id));
-    } else {
-      console.error("Delete failed:", res.error);
+    if (result.isConfirmed) {
+      setLoader(true);
+
+      const res = await deleteData({
+        url: `/dashboard/projects/api?id=${id}`,
+        setLoader,
+      });
+
+      if (!res.error) {
+        setProjects((prev) => prev.filter((blog) => blog._id !== id));
+        Swal.fire("Deleted!", "Your item has been deleted.", "success");
+      } else {
+        Swal.fire("Error", "Something went wrong while deleting.", "error");
+        console.error("Delete failed:", res.error);
+      }
+
+      setLoader(false);
     }
   };
 
-  // console.log(loader, projects, "check-loader");
   return (
     <Table className="overflow-y-auto h-fit">
       <TableHeader>
