+++ src/components/team/TeamTable.tsx
@@ -12,6 +12,7 @@ import {
 } from "../ui/table";
 import { DeleteIcon } from "lucide-react";
 import Image from "next/image";
+import Swal from "sweetalert2";
 type TeamItem = {
   _id: string;
   name: string;
@@ -49,20 +50,36 @@ const TeamTable = () => {
 
     fetchData();
   }, []);
+
   const handleDelete = async (id: string) => {
-    const res = await deleteData({
-      url: `/dashboard/team/api?id=${id}`,
-      setLoader,
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to undo this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "Cancel",
     });
 
-    if (!res.error) {
-      setTeam((prev) => prev.filter((team) => team._id !== id));
-    } else {
-      console.error("Delete failed:", res.error);
+    if (result.isConfirmed) {
+      setLoader(true);
+
+      const res = await deleteData({
+        url: `/dashboard/team/api?id=${id}`,
+        setLoader,
+      });
+
+      if (!res.error) {
+        setTeam((prev) => prev.filter((blog) => blog._id !== id));
+        Swal.fire("Deleted!", "Your item has been deleted.", "success");
+      } else {
+        Swal.fire("Error", "Something went wrong while deleting.", "error");
+        console.error("Delete failed:", res.error);
+      }
+
+      setLoader(false);
     }
   };
-
-  // console.log(loader, teams, "check-loder, teams");
   return (
     <Table className="overflow-y-auto max-h-[400px]">
       <TableHeader>
